fix: id : String -> Long
속도, 인덱스 크기, 관리 용이

+++ src/main/java/techtrek/domain/interviewQuestion/repository/InterviewQuestionRepository.java
@@ -9,11 +9,11 @@ import techtrek.domain.interviewQuestion.entity.InterviewQuestion;
 import java.util.Optional;
 
 @Repository
-public interface InterviewQuestionRepository extends JpaRepository<InterviewQuestion, Integer> {
+public interface InterviewQuestionRepository extends JpaRepository<InterviewQuestion, Long> {
     boolean existsByQuestion(String question);
 
     // 특정 기업의 질문 중 랜덤 1개 가져오기
     @Query(value = "SELECT * FROM interview_question WHERE enterprise_id = :enterpriseId ORDER BY RAND() LIMIT 1", nativeQuery = true)
-    Optional<InterviewQuestion> findRandomQuestionByEnterpriseId(@Param("enterpriseId") int enterpriseId);
+    Optional<InterviewQuestion> findRandomQuestionByEnterpriseId(@Param("enterpriseId") Long enterpriseId);
 
 }
